extracted logging to a separate module
