+++ executor-service/build.gradle.kts
@@ -1,7 +1,6 @@
 dependencies {
     implementation("com.squareup.okhttp3:okhttp:4.11.0")
     implementation("org.seleniumhq.selenium:selenium-java:4.13.0")
-    implementation("org.aspectj:aspectjweaver:1.9.20.1")
     implementation("com.fasterxml.jackson.core:jackson-databind:2.15.2")
-    implementation(project(":logback"))
+    implementation(project(":logger"))
 }
