+++ logger/build.gradle.kts
@@ -1,0 +1,8 @@
+dependencies {
+    implementation("org.aspectj:aspectjweaver:1.9.20.1")
+    implementation(project(":logback"))
+}
+
+tasks.bootJar {
+    enabled = false
+}
