+++ logger/src/main/java/executor/service/annotation/HandleException.java
@@ -1,0 +1,11 @@
+package executor.service.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface HandleException {
+}
