+++ logger/src/main/java/executor/service/logger/ExceptionHandlingAspect.java
@@ -1,6 +1,5 @@
 package executor.service.logger;
 
-import executor.service.exception.scenario.ScenarioExecutionException;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
@@ -11,14 +10,14 @@ import org.springframework.stereotype.Component;
 @Component
 @Aspect
 @RequiredArgsConstructor
-public class ScenarioExceptionHandlingAspect {
+public class ExceptionHandlingAspect {
     private final Logger logger;
 
-    @Around("execution(* executor.service.execution.scenario.ScenarioExecutor+.*(..))")
+    @Around("@within(executor.service.annotation.HandleException)")
     public void handle(ProceedingJoinPoint joinPoint) throws Throwable {
         try {
             joinPoint.proceed();
-        } catch (ScenarioExecutionException ex) {
+        } catch (Exception ex) {
             String className = joinPoint.getTarget().getClass().getSimpleName();
             String methodName = joinPoint.getSignature().getName();
             logger.error(LogMessage.INVOCATION_TARGET_EXCEPTION.getMessage(), methodName, className, ex);
