+++ logger/src/main/java/executor/service/logger/LogMessage.java
