+++ settings.gradle.kts
@@ -1,3 +1,4 @@
 rootProject.name = "worker_microservice"
 include("executor-service")
 include("logback")
+include("logger")
